[SPE-158] Handle Json Value when reading decryption body response (#115)
* [SPE-158] Handle more than Json Value when reading decryption body response

* [SPE-158] Add more cases to the tests

* update parser — remove irrelevant fields being captured, introduce contextful ciphertext parsing

* update stream to handle optional final tokens so we can track quotes. TODO: add tests for stripping quotes on non string encrypted data types

* Stream now correctly mapping ciphertexts to ranges including their quotes if non-string types

* format

* lint: remove unused imports

* add comment to stream tests

* update customer process to respond with 500 & error, only run kill command when not in CI

* update tests - parse data types in mock crypto api

* add rust crypto mock api to replace existing node api

* update workflow to pass registry info from secrets

* restore daemon flag for docker run args

* update attestation doc endpoint to not require json header, update test runner to provide cargo registry via github secrets

* update readme with instructions for setting up the e2e tests locally

* add instructions for cross compiler and add target

* add target explicitly

* target added in wrong code path

* add target correctly

* update e2e tests script to not compile in CI, use static compilation workflow in action

* add error logs to e2e tests, don't kill container after testing

* increase wait in tests, customer process isn't starting quickly enough to respond to tests

* chmod mock crypto api

* ty @nuggimane, was incorrectly reading in certs

* update tests to evaluate data plane in flight decryption

* update parser to match debug encrypted strings

* format

* add in support for debug ciphertext parsing

* update tests to decrypt data using the data plane, as well as explicit calls to the crypto api

* format

* remove logs from debugging decrypts

* extend e2e tests to include nested objects and arrays

* fix json parsing tests

+++ e2e-tests/run.sh
@@ -1,7 +1,10 @@
 set -e 
 
 # kill container if it is left running by hanging test
-docker kill cages-test-container || true
+if [[ -z "${CI}" ]];
+then
+  docker kill cages-test-container || true
+fi
 
 # install the node modules for customer process and test script
 cd e2e-tests && npm install && cd ..
@@ -19,7 +22,7 @@ if [[ -z "${CI}" ]];
 then
   cargo build --features network_egress --release --target x86_64-unknown-linux-musl
   echo "Building cage container"
-  docker build --no-cache --platform=linux/amd64 --build-arg MOCK_CRYPTO_CERT=/services/testing.crt --build-arg MOCK_CRYPTO_KEY=/services/testing.key -f e2e-tests/Dockerfile -t cages-test .
+  docker build --platform=linux/amd64 --build-arg MOCK_CRYPTO_CERT=/services/testing.crt --build-arg MOCK_CRYPTO_KEY=/services/testing.key -f e2e-tests/Dockerfile -t cages-test .
 else
   echo "Building cage container CI"
   docker build --build-arg=CI=true --build-arg MOCK_CRYPTO_CERT="$MOCK_CRYPTO_CERT" --build-arg MOCK_CRYPTO_KEY="$MOCK_CRYPTO_KEY" --platform=linux/amd64 -f e2e-tests/Dockerfile -t cages-test .
